Example using data
A simple example of the Module being used. I took a data set I found on the internet (Mushrooms with poisonous ratings), I converted the poison factor into a binary 1/0. I also normalised the data. I then trained my Neural Network on it, and outputted the ROC curves for Training and Test. As can be seen, the ROC curves show positive predictability  for both training and test results